esto no se acaba nunca. debugando casos difíciles salen lagunas que provocan la inserción de bastante más código. continuaré con el debug-design-debug.

+++ trunk/mastermind/src/CPUAlgorithm/MastermindAI4Holes.java
@@ -67,7 +67,7 @@ public class MastermindAI4Holes extends CommonMastermindAI
     protected void SuccessorSelection(Integer lastRow, Integer[] successor)
     {
 
-        if (lastRow >= -1 && lastRow <= 2) {
+        if (lastRow.intValue() >= -1 && lastRow.intValue() <= 2) {
 
             switch (lastRow) {
                 case -1:
